param: Fold mgt_param_bits.c into mgt_param_tweak.c
