+++ bin/varnishd/Makefile.am
@@ -85,7 +85,6 @@ varnishd_SOURCES = \
 	mgt/mgt_jail_unix.c \
 	mgt/mgt_main.c \
 	mgt/mgt_param.c \
-	mgt/mgt_param_bits.c \
 	mgt/mgt_param_tcp.c \
 	mgt/mgt_param_tweak.c \
 	mgt/mgt_shmem.c \
